added domains to load_omop, renamed setup_omop.py ddl.py, fixed DDL for another bigint, fixed config spec.s

+++ resources/OMOPCDM_duckdb_5.3_ddl_with_constraints_and_bigint_PK.sql
@@ -306,7 +306,7 @@ CREATE TABLE @cdmDatabaseSchema.care_site (
 			care_site_id bigint NOT NULL,
 			care_site_name varchar(255) NULL,
 			place_of_service_concept_id integer NULL,
-			location_id integer NULL,
+			location_id bigint NULL,
 			care_site_source_value varchar(50) NULL,
 			place_of_service_source_value varchar(50) NULL );
 --HINT DISTRIBUTE ON RANDOM
